requirements.txt
# Core ML and LLM training dependencies
torch>=2.1.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0

# Data processing and utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
pyyaml>=6.0

# Jupyter and visualization
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.15.0

# Optional: Additional utilities
tqdm>=4.65.0
psutil>=5.9.0
pandas>=2.0.0
tokenizers>=0.14.0
scikit-learn>=1.3.0

# Training and evaluation
wandb>=0.15.0
tensorboard>=2.14.0
evaluate>=0.4.0

# Development and notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress tracking and utilities
tqdm>=4.65.0
rich>=13.0.0

# Optional: For specific model architectures
# sentence-transformers>=2.2.0
# flash-attn>=2.3.0  # For faster attention (if compatible with Apple Silicon)
